+++ sdk/openai/openai/src/models/options.ts
@@ -233,4 +233,14 @@ export interface GetChatCompletionsOptions extends OperationOptions {
    *   This additional specification is only compatible with Azure OpenAI.
    */
   azureExtensionOptions?: AzureExtensionsOptions;
+  /**
+   * Whether to return log probabilities of the output tokens or not.If true, returns the log probabilities of
+   * each output token returned in the `content` of `message`. This option is currently not available on the `gpt-4-vision-preview` model.
+   */
+  logprobs?: boolean;
+  /**
+   * An integer between 0 and 5 specifying the number of most likely tokens to return at each token position,
+   * each with an associated log probability. `logprobs` must be set to `true` if this parameter is used.
+   */
+  topLogprobs?: number;
 }